+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.republikapp",
+    package = "app.republik",
 )
 
 android_resource(
     name = "res",
-    package = "com.republikapp",
+    package = "app.republik",
     res = "src/main/res",
 )
