+++ android/app/src/main/java/app/republik/MainApplication.java
@@ -1,18 +1,17 @@
-package com.republikapp;
+package app.republik;
 
 import android.app.Application;
 import android.content.Context;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
-import com.guichaguri.trackplayer.TrackPlayer;
 import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
-import org.devio.rn.splashscreen.SplashScreenReactPackage;
-import com.wix.reactnativenotifications.RNNotificationsPackage;
+
+import app.republik.BuildConfig;
 
 public class MainApplication extends Application implements ReactApplication {
 
@@ -65,7 +64,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.republikapp.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("app.republik.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
